Restructure project sources
