+++ include/reversi/player.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <string>
+
+class Player {
+public:
+
+private:
+  std::string m_name;
+};
