+++ include/reversi/utils.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <iostream>
+#include <vector>
+
+template<class T>
+std::ostream& operator<<(std::ostream& os, const std::vector<T>& vector) {
+  for (const T& item : vector) {
+    os << item << ", ";
+  }
+  return os;
+}
